+++ weblate/api/serializers.py
@@ -364,7 +364,7 @@ class UploadRequestSerializer(ReadOnlySerializer):
     author_email = serializers.EmailField(required=False)
     author_name = serializers.CharField(max_length=200, required=False)
     method = serializers.ChoiceField(
-        choices=("translate", "approve", "suggest", "fuzzy", "replace"),
+        choices=("translate", "approve", "suggest", "fuzzy", "replace", "source"),
         required=False,
         default="translate",
     )